rschir/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
matplotlib==3.7.4
PyMySQL==1.1.0